Live: /auth/signup with parameter and database validation

+++ src/main/java/com/preplane/dev/security/services/UserDetailsImpl.java
@@ -15,7 +15,7 @@ public class UserDetailsImpl implements UserDetails {
     // Interal versioning
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private int id;
     private String username;
     private String email;
     @JsonIgnore
@@ -23,7 +23,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String username, String email, String password,
+    public UserDetailsImpl(int id, String username, String email, String password,
             Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
@@ -36,7 +36,7 @@ public class UserDetailsImpl implements UserDetails {
         List<GrantedAuthority> authorities = List.of(new SimpleGrantedAuthority(user.getRole()));
 
         return new UserDetailsImpl(
-                Long.valueOf(user.getId()),
+                user.getId(),
                 user.getUsername(),
                 user.getEmailAddress(),
                 user.getPassword(),
@@ -48,7 +48,7 @@ public class UserDetailsImpl implements UserDetails {
         return authorities;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
